requirements.txt
pydantic-ai
openai
python-dotenv
fastapi
httpx
uvicorn
